Add Vcf and Fasta line parsers.
These parsers will reduce unnecessary memory usage by reading the files
line-by-line and yielding contig positions instead of reading the entire
file into memory.
Update matrix_dto to return Vcf and Fasta objects. At initialization the objects
build an index of the starting file position of each contig in the file.
The purpose of the index will be to process the contigs in parallel.